Packets Work With Any Transport Now, and EzWriter Was Quietly Eating Bytes Past 64
-EzWriter Wasnt Surviving Its Own Buffer Growth
- IDs Are 32-bit Now, Collisions Basically Cant Happen
- Dispatch is Per Endpoint Now, Static Was a Mistake
- Peers Belong to the Transport, Not to - EzMultiLib Core Library Doesnt Drag in Roslyn Anymore
- Shoutout Claude For catching the static mistake, wouldn't have thought of it for a while

+++ EzMultiLib/EzMultiLib/Packets/PacketIdAttribute.cs
@@ -5,9 +5,9 @@ namespace EzMultiLib.Packets
 	[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, AllowMultiple = false, Inherited = false)]
 	public sealed class PacketIdAttribute : Attribute
 	{
-		public ushort Id { get; }
+		public uint Id { get; }
 
-		public PacketIdAttribute(ushort id)
+		public PacketIdAttribute(uint id)
 		{
 			Id = id;
 		}
